backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# ML Integration (simplified - no compiled dependencies)
# joblib==1.3.2
# scikit-learn==1.3.2
# pandas==2.1.3
# numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
# groq==0.11.0  # REMOVED - Using direct HTTP requests instead to avoid 'proxies' parameter error

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Redis (optional)
redis==5.0.1
